Routing: Fix encoding.
When using encoding, some characters got encoded a few times, so the
upstream request was not correct.

Is impossible to get it 100% perfect, because some chars (`?/#`) are
hard to get it right when using ngx.var.uri, but we need to use
ngx.var.uri.

After discussing it with support/product, this is the best way to handle
it.

Reported by sillumin@redhat.com
Fix: https://issues.redhat.com/browse/THREESCALE-5117

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>