disable request buffering when upstream in plain HTTP 1.1
It is required to not read request body for chunked requests to be propagated as chunked requests.
Only read the request body when POST query arguments (of the MIME type application/x-www-form-urlencoded)
and thus query arguments could be in the request body.

Note: chunked requests are propagated to upstream with Transfer-Encoding chunked only when client request sends more than one chunk