[configuration] Reset configuration cache on reload
When reloading the configuration, APIcast only updates the services
that present in the new configuration and leaves the old/deleted services
in the cache. When a new request is sent to the deleted service, APIcast
will reuse the stale configuration in the cache, leading to unexpected
behavior.

This PR ensures that APIcast always resets the cache when reloading
the configuration